prevent deletion of HoldingsRecordsSources associated with HoldingsRecords

+++ src/main/java/org/folio/rest/impl/HoldingsRecordsSourceAPI.java
@@ -2,14 +2,17 @@ package org.folio.rest.impl;
 
 import static io.vertx.core.Future.succeededFuture;
 
+import java.util.List;
 import java.util.Map;
 
 import javax.ws.rs.core.Response;
 
 import org.folio.cql2pgjson.CQL2PgJSON;
 import org.folio.cql2pgjson.exception.FieldException;
+import org.folio.rest.jaxrs.model.HoldingsRecord;
 import org.folio.rest.jaxrs.model.HoldingsRecordsSource;
 import org.folio.rest.jaxrs.model.HoldingsRecordsSources;
+import org.folio.rest.jaxrs.resource.HoldingsStorage.GetHoldingsStorageHoldingsByHoldingsRecordIdResponse;
 import org.folio.rest.persist.PgUtil;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.persist.Criteria.Limit;
@@ -22,6 +25,7 @@ import org.z3950.zing.cql.CQLParseException;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
+import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
@@ -29,6 +33,7 @@ import io.vertx.core.logging.LoggerFactory;
 public class HoldingsRecordsSourceAPI implements org.folio.rest.jaxrs.resource.HoldingsRecordsSources {
 
   private static final String REFERENCE_TABLE = "holdings_records_source";
+  public static final String HOLDINGS_RECORD_TABLE = "holdings_record";
   private static final Logger log = LoggerFactory.getLogger(HoldingsRecordsSourceAPI.class);
   private final Messages messages = Messages.getInstance();
 
@@ -69,13 +74,45 @@ public class HoldingsRecordsSourceAPI implements org.folio.rest.jaxrs.resource.H
     vertxContext.runOnContext(v -> {
       try {
         String tenantId = TenantTool.tenantId(okapiHeaders);
-        PostgresClient.getInstance(vertxContext.owner(), tenantId).getById(REFERENCE_TABLE,
+        PostgresClient pgClient = PostgresClient.getInstance(vertxContext.owner(), tenantId);
+        pgClient.getById(REFERENCE_TABLE,
             id, HoldingsRecordsSource.class,
             reply -> {
               if (reply.succeeded()) {
                 String source = reply.result().getSource();
-                if (!source.contentEquals("folio") && !source.contentEquals("marc")) {
-                  PgUtil.deleteById(REFERENCE_TABLE, id, okapiHeaders, vertxContext, DeleteHoldingsRecordsSourcesByIdResponse.class, asyncResultHandler);
+                if (source != null && !source.contentEquals("folio") && !source.contentEquals("marc")) {
+                  try {
+                    String[] fieldList = {"*"};
+                    CQLWrapper cqlWrapper = getCQL(HOLDINGS_RECORD_TABLE, String.format("sourceId==%s", id), 1, 0);
+                    pgClient.get(HOLDINGS_RECORD_TABLE, HoldingsRecord.class, fieldList, cqlWrapper.getQuery(), true, false,
+                        holdingsRecordReply -> {
+                          try {
+                            if (holdingsRecordReply.succeeded()) {
+                              List<HoldingsRecord> holdingsList = holdingsRecordReply.result().getResults();
+                              if (holdingsList.size() == 0) {
+                                PgUtil.deleteById(REFERENCE_TABLE, id, okapiHeaders, vertxContext, DeleteHoldingsRecordsSourcesByIdResponse.class, asyncResultHandler);
+                              } else {
+                                log.error("Holdings Records Sources associated with Holdings Record(s) can not be deleted");
+                                asyncResultHandler.handle(succeededFuture(GetHoldingsRecordsSourcesResponse
+                                  .respond400WithTextPlain("Holdings Records Sources associated with Holdings Record(s) can not be deleted")));
+                              }
+                            } else {
+                              asyncResultHandler.handle(
+                                Future.succeededFuture(
+                                  GetHoldingsStorageHoldingsByHoldingsRecordIdResponse.
+                                    respond500WithTextPlain(holdingsRecordReply.cause().getMessage())));
+                            }
+                          } catch (Exception e) {
+                              log.error(e.getMessage());
+                            asyncResultHandler.handle(io.vertx.core.Future.succeededFuture(
+                              GetHoldingsStorageHoldingsByHoldingsRecordIdResponse.
+                                respond500WithTextPlain(e.getMessage())));
+                          }
+                        });
+
+                  } catch (Exception e) {
+
+                  }
                 } else {
                   log.error("Holdings Records Sources with source of folio or marc can not be deleted");
                   asyncResultHandler.handle(succeededFuture(GetHoldingsRecordsSourcesResponse
@@ -103,7 +140,11 @@ public class HoldingsRecordsSourceAPI implements org.folio.rest.jaxrs.resource.H
   }
 
   private CQLWrapper getCQL(String query, int limit, int offset) throws FieldException {
-    CQL2PgJSON cql2pgJson = new CQL2PgJSON(REFERENCE_TABLE+".jsonb");
+    return getCQL(REFERENCE_TABLE, query, limit, offset);
+  }
+
+  private CQLWrapper getCQL(String table, String query, int limit, int offset) throws FieldException {
+    CQL2PgJSON cql2pgJson = new CQL2PgJSON(table+".jsonb");
     return new CQLWrapper(cql2pgJson, query).setLimit(new Limit(limit)).setOffset(new Offset(offset));
   }
 }
